refactor: split gateway types into types.rs, remove dead code
- Extract DmScope, Message, GatewayConfig, Session, InboundChainInput,
  GatewayError, OutputTx into crates/gateway/src/types.rs
- Add pub mod types + pub use types::* in lib.rs
- Remove #[allow(dead_code)] from 4 methods (has_slash_dispatcher,
  has_session_handler, config_name, processor_registry_len) which are
  called from tests
- Delete src/gateway_adapters.rs (dead code file, not declared as module)
- lib.rs reduced from 1000 to 831 lines

Source: issue #1252
Type: refactor
